Added some more authentication.

+++ routers/order.router.ts
@@ -1,23 +1,35 @@
 import {Router} from "express";
 import {handleFindVariables} from "../utils/validation";
 import {OrderRecord} from "../records/order.record";
-import {OrderStatus} from "../types";
-import {ValidationError} from "../utils/error";
+import {OrderStatus, UserType} from "../types";
+import {AccessError, ValidationError} from "../utils/error";
+import {AuthenticationRequest} from "../types/auth/auth";
+import {authenticate} from "../utils/auth";
+import {UserRecord} from "../records/user.record";
 
 export const orderRouter = Router()
-    .get('/', async (req, res) => {
+    .get('/', authenticate, async (req: AuthenticationRequest, res) => {
+        const user = await UserRecord.getOne(req.user.id, true);
+        if (user.userType === UserType.Production) throw new AccessError("Forbidden");
+
         const {sortAsc, name, rows, page} = handleFindVariables(req.body)
         const tasks = await OrderRecord.find(sortAsc, name, rows, page);
         res.json(tasks);
     })
-    .post('/', async (req, res) => {
+    .post('/', authenticate, async (req: AuthenticationRequest, res) => {
+        const user = await UserRecord.getOne(req.user.id, true);
+        if (user.userType === UserType.Production || user.userType === UserType.Warehouse) throw new AccessError("Forbidden");
+
         const order = new OrderRecord(req.body);
         await order.insert();
         res
             .status(201)
             .json(order);
     })
-    .patch('/', async (req, res) => {
+    .patch('/', authenticate, async (req: AuthenticationRequest, res) => {
+        const user = await UserRecord.getOne(req.user.id, true);
+        if (user.userType === UserType.Production) throw new AccessError("Forbidden");
+
         const order = await OrderRecord.getOne(req.body.id);
         const {name, description, status} = req.body;
         let changed = false;
@@ -46,7 +58,10 @@ export const orderRouter = Router()
             .status(202)
             .json(order);
     })
-    .delete('/', async (req, res) => {
+    .delete('/', authenticate, async (req: AuthenticationRequest, res) => {
+        const user = await UserRecord.getOne(req.user.id, true);
+        if (user.userType !== UserType.Admin) throw new AccessError("Forbidden");
+
         const order = await OrderRecord.getOne(req.body.id);
         await order.delete();
         res
